Debugger modal (#147)
Debugger modal for helping user understand how dynamic_states and args interact.
Shows a view that prints out inputs and states from the variableselector, and also has the option to pass a function which will get its return value printed.
It is meant as a helper view to figure out how the framework is put together.